chore: conflict 해결 및 log 로직 분리
