+++ src/main/java/com/example/jomajomadelivery/orders/service/OrdersService.java
@@ -16,9 +16,7 @@ import com.example.jomajomadelivery.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.coyote.BadRequestException;
-import org.aspectj.lang.annotation.AfterReturning;
 import org.aspectj.lang.annotation.Aspect;
-import org.aspectj.lang.annotation.Pointcut;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,21 +32,6 @@ public class OrdersService {
     private final OrdersRepository ordersRepository;
     private final UserRepository userRepository;
     private final CartRepository cartRepository;
-
-    @Pointcut("execution(* com.example.jomajomadelivery.orders.service.OrdersService.create(..)) ||" +
-            "execution(* com.example.jomajomadelivery.orders.service.OrdersService.update(..))")
-    public void ordersServicePointcut() {}
-
-    @AfterReturning(pointcut = "ordersServicePointcut()",returning = "result")
-    public void logOrderActivity(Object result) {
-        if (result instanceof Order order) {
-            log.info("Order Log - 상태: {}, 요청 시각: {}, 가게 ID: {}, 주문 ID: {}",
-                    order.getStatus(),
-                    LocalTime.now(),
-                    order.getStore().getStoreId(),
-                    order.getOrderId());
-        }
-    }
 
     //Todo:: User, Store, Cart, Address 주입 필요
     public OrderResponseDto create() {
